requirements-test.txt
django-allauth
djangorestframework
pytest
pytest-cov
pytest-django
pytest-mock
responses
social-auth-app-django
